backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
numpy==1.26.0
scikit-learn==1.3.2
joblib==1.3.2
werkzeug==2.3.7
tensorflow==2.8.0
pandas==1.4.0
flask_sqlalchemy
Pillow
pickle-mixin